+++ tests/tpow.c
@@ -786,7 +786,8 @@ particular_cases (void)
           {
             printf ("Error in mpfr_pow for (%s)^(%s) (%d,%d):\n"
                     "Flags = %u instead of expected %u\n",
-                    name[i], name[j], i, j, __gmpfr_flags, f[i][j]);
+                    name[i], name[j], i, j,
+                    (unsigned int) __gmpfr_flags, f[i][j]);
             mpfr_dump (r);
             error = 1;
           }
@@ -815,7 +816,8 @@ particular_cases (void)
               {
                 printf ("Error in mpfr_pow_z for (%s)^(%s) (%d,%d):\n"
                         "Flags = %u instead of expected %u\n",
-                        name[i], name[j], i, j, __gmpfr_flags, f[i][j]);
+                        name[i], name[j], i, j,
+                        (unsigned int) __gmpfr_flags, f[i][j]);
                 mpfr_dump (r);
                 error = 1;
               }
@@ -840,7 +842,8 @@ particular_cases (void)
               {
                 printf ("Error in mpfr_pow_si for (%s)^(%s) (%d,%d):\n"
                         "Flags = %u instead of expected %u\n",
-                        name[i], name[j], i, j, __gmpfr_flags, f[i][j]);
+                        name[i], name[j], i, j,
+                        (unsigned int) __gmpfr_flags, f[i][j]);
                 mpfr_dump (r);
                 error = 1;
               }
@@ -866,7 +869,8 @@ particular_cases (void)
                   {
                     printf ("Error in mpfr_pow_ui for (%s)^(%s) (%d,%d):\n"
                             "Flags = %u instead of expected %u\n",
-                            name[i], name[j], i, j, __gmpfr_flags, f[i][j]);
+                            name[i], name[j], i, j,
+                            (unsigned int) __gmpfr_flags, f[i][j]);
                     mpfr_dump (r);
                     error = 1;
                   }
@@ -896,7 +900,8 @@ particular_cases (void)
               {
                 printf ("Error in mpfr_ui_pow for (%s)^(%s) (%d,%d):\n"
                         "Flags = %u instead of expected %u\n",
-                        name[i], name[j], i, j, __gmpfr_flags, f[i][j]);
+                        name[i], name[j], i, j,
+                        (unsigned int) __gmpfr_flags, f[i][j]);
                 mpfr_dump (r);
                 error = 1;
               }
@@ -1106,7 +1111,7 @@ overflows2 (void)
       if (__gmpfr_flags != (MPFR_FLAGS_OVERFLOW | MPFR_FLAGS_INEXACT))
         {
           printf ("Error in overflows2 (e = %d): bad flags (%u)\n",
-                  e, __gmpfr_flags);
+                  e, (unsigned int) __gmpfr_flags);
           exit (1);
         }
     }
@@ -1155,7 +1160,7 @@ overflows3 (void)
           if (__gmpfr_flags != (MPFR_FLAGS_OVERFLOW | MPFR_FLAGS_INEXACT))
             {
               printf ("Error in overflows3 (RNDN, i = %d): bad flags (%u)\n",
-                      i, __gmpfr_flags);
+                      i, (unsigned int) __gmpfr_flags);
               exit (1);
             }
 
@@ -1285,7 +1290,8 @@ bug20071104 (void)
     }
   if (__gmpfr_flags != (MPFR_FLAGS_OVERFLOW | MPFR_FLAGS_INEXACT))
     {
-      printf ("Error in bug20071104: bad flags (%u)\n", __gmpfr_flags);
+      printf ("Error in bug20071104: bad flags (%u)\n",
+              (unsigned int) __gmpfr_flags);
       exit (1);
     }
   mpfr_clears (x, y, z, (mpfr_ptr) 0);
